# Use bools rather than bit fields for DCEL labels