--- a/geom/dcel_extract_intersection_matrix.go
+++ b/geom/dcel_extract_intersection_matrix.go
@@ -3,50 +3,54 @@ package geom
 func (d *doublyConnectedEdgeList) extractIntersectionMatrix() matrix {
 	im := newMatrix()
 	for _, v := range d.vertices {
-		locA := v.location(inputAMask)
-		locB := v.location(inputBMask)
+		locA := v.location(operandA)
+		locB := v.location(operandB)
 		im.set(locA, locB, '0')
 	}
 	for _, e := range d.halfEdges {
-		locA := e.location(inputAMask)
-		locB := e.location(inputBMask)
+		locA := e.location(operandA)
+		locB := e.location(operandB)
 		im.set(locA, locB, '1')
 	}
 	for _, f := range d.faces {
-		locA := f.location(inputAMask)
-		locB := f.location(inputBMask)
+		locA := f.location(operandA)
+		locB := f.location(operandB)
 		im.set(locA, locB, '2')
 	}
 	return im
 }
 
-func (f *faceRecord) location(sideMask uint8) imLocation {
-	if (f.label & inSetMask & sideMask) == 0 {
+func (f *faceRecord) location(operand operand) imLocation {
+	// TODO: Should we assert isPopulated?
+	if !f.labels[operand].inSet {
 		return imExterior
 	}
 	return imInterior
 }
 
-func (e *halfEdgeRecord) location(sideMask uint8) imLocation {
-	if (e.edgeLabel & inSetMask & sideMask) == 0 {
+func (e *halfEdgeRecord) location(operand operand) imLocation {
+	// TODO: Should we assert isPopulated?
+	if !e.edgeLabels[operand].inSet {
 		return imExterior
 	}
-	face1Present := (e.incident.label & inSetMask & sideMask) != 0
-	face2Present := (e.twin.incident.label & inSetMask & sideMask) != 0
+	face1Present := e.incident.labels[operand].inSet
+	face2Present := e.twin.incident.labels[operand].inSet
 	if face1Present != face2Present {
 		return imBoundary
 	}
 	return imInterior
 }
 
-func (v *vertexRecord) location(sideMask uint8) imLocation {
+func (v *vertexRecord) location(operand operand) imLocation {
+	// TODO: Should we assert isPopulated?
+
 	// NOTE: It's important that we check the Boundary flag before the Interior
 	// flag, since both might be set. In that case, we want to treat the
 	// location as a Boundary, since the boundary is a more specific case.
 	switch {
-	case (v.locLabel & sideMask & locBoundary) != 0:
+	case v.locations[operand].boundary:
 		return imBoundary
-	case (v.locLabel & sideMask & locInterior) != 0:
+	case v.locations[operand].interior:
 		return imInterior
 	default:
 		// We don't know the location of the point. But it must be either
@@ -54,7 +58,7 @@ func (v *vertexRecord) location(sideMask uint8) imLocation {
 		// that. We can just use the location of one of the incident edges,
 		// since that would have the same location.
 		for _, e := range v.incidents {
-			return e.location(sideMask)
+			return e.location(operand)
 		}
 		panic("point has no incidents") // Can't happen, due to ghost edges.
 	}