… geom/dcel_overlay.go
… geom/dcel_test.go