## Description

Simplifies code by using explicit bools rather than bit fields.

## Check List

Have you:

- Added unit tests? N/A, existing.

- Add cmprefimpl tests? (if appropriate?) N/A, existing.

## Related Issue

- https://github.com/peterstace/simplefeatures/issues/341

## Benchmark Results

No significant change:

<details>

<summary>Click to expand</summary>

```
name                                                      old time/op    new time/op    delta
pkg:github.com/peterstace/simplefeatures/geom goos:linux goarch:amd64
LineEnvelope/0                                              2.34ns ± 8%    1.96ns ± 5%  -16.29%  (p=0.000 n=15+15)
LineEnvelope/1                                              1.37ns ± 4%    1.35ns ± 3%     ~     (p=0.241 n=14+14)
LineEnvelope/2                                              1.70ns ± 3%    1.69ns ± 3%     ~     (p=0.454 n=14+15)
LineEnvelope/3                                              1.55ns ± 3%    1.52ns ± 2%   -1.64%  (p=0.010 n=15+15)
MarshalWKB/polygon/n=10                                      231ns ± 4%     234ns ± 6%     ~     (p=0.124 n=14+15)
MarshalWKB/polygon/n=100                                     454ns ± 6%     453ns ± 6%     ~     (p=0.583 n=14+15)
MarshalWKB/polygon/n=1000                                   2.49µs ± 4%    2.53µs ± 8%     ~     (p=0.400 n=14+15)
MarshalWKB/polygon/n=10000                                  21.4µs ± 4%    21.4µs ± 6%     ~     (p=1.000 n=14+14)
UnmarshalWKB/polygon/n=10                                    408ns ± 8%     406ns ± 4%     ~     (p=0.919 n=14+14)
UnmarshalWKB/polygon/n=100                                   631ns ± 2%     629ns ± 6%     ~     (p=0.355 n=13+13)
UnmarshalWKB/polygon/n=1000                                 2.88µs ± 4%    2.93µs ± 3%   +1.75%  (p=0.012 n=13+14)
UnmarshalWKB/polygon/n=10000                                28.8µs ± 4%    29.0µs ± 5%     ~     (p=0.525 n=15+13)
IntersectsLineStringWithLineString/n=10                     1.72µs ± 9%    1.71µs ± 2%     ~     (p=0.890 n=13+13)
IntersectsLineStringWithLineString/n=100                    24.5µs ± 9%    24.7µs ± 5%     ~     (p=0.158 n=14+15)
IntersectsLineStringWithLineString/n=1000                    255µs ± 3%     258µs ± 6%     ~     (p=0.621 n=14+15)
IntersectsLineStringWithLineString/n=10000                  3.40ms ± 7%    3.44ms ± 5%     ~     (p=0.194 n=14+14)
IntersectsMultiPointWithMultiPoint/n=20                     1.85µs ± 2%    1.85µs ± 2%     ~     (p=0.675 n=15+15)
IntersectsMultiPointWithMultiPoint/n=200                    19.8µs ± 4%    19.8µs ± 4%     ~     (p=0.914 n=14+15)
IntersectsMultiPointWithMultiPoint/n=2000                    194µs ± 5%     196µs ± 4%     ~     (p=0.187 n=15+15)
IntersectsMultiPointWithMultiPoint/n=20000                  2.08ms ± 3%    2.13ms ± 3%   +2.60%  (p=0.001 n=12+15)
PolygonSingleRingValidation/n=10                            3.05µs ± 6%    3.06µs ± 3%     ~     (p=0.367 n=15+15)
PolygonSingleRingValidation/n=100                           43.6µs ± 4%    43.6µs ± 3%     ~     (p=0.940 n=15+14)
PolygonSingleRingValidation/n=1000                           526µs ± 6%     518µs ± 2%   -1.49%  (p=0.041 n=15+13)
PolygonSingleRingValidation/n=10000                         6.96ms ± 4%    6.96ms ± 4%     ~     (p=0.914 n=14+15)
PolygonMultipleRingsValidation/n=4                          8.58µs ± 4%    8.55µs ± 3%     ~     (p=0.614 n=13+13)
PolygonMultipleRingsValidation/n=36                         72.8µs ± 3%    73.0µs ± 3%     ~     (p=0.430 n=14+13)
PolygonMultipleRingsValidation/n=400                         961µs ± 5%     977µs ± 6%     ~     (p=0.239 n=14+13)
PolygonMultipleRingsValidation/n=4096                       12.3ms ± 3%    12.1ms ± 4%     ~     (p=0.477 n=14+15)
PolygonZigZagRingsValidation/n=10                           14.3µs ± 4%    14.2µs ± 3%     ~     (p=0.351 n=15+15)
PolygonZigZagRingsValidation/n=100                           167µs ± 4%     164µs ± 3%   -1.63%  (p=0.010 n=15+12)
PolygonZigZagRingsValidation/n=1000                         1.97ms ± 2%    1.98ms ± 3%     ~     (p=0.541 n=14+14)
PolygonZigZagRingsValidation/n=10000                        26.5ms ± 5%    26.5ms ± 7%     ~     (p=0.793 n=13+14)
PolygonAnnulusValidation/n=10                               4.54µs ± 4%    4.56µs ± 3%     ~     (p=0.344 n=13+14)
PolygonAnnulusValidation/n=100                              41.4µs ± 4%    42.2µs ± 5%   +1.98%  (p=0.037 n=14+15)
PolygonAnnulusValidation/n=1000                              730µs ± 5%     724µs ± 4%     ~     (p=0.285 n=15+15)
PolygonAnnulusValidation/n=10000                            8.49ms ± 3%    8.55ms ± 3%     ~     (p=0.285 n=15+15)
MultipolygonValidation/n=1                                   485ns ± 2%     487ns ± 3%     ~     (p=0.419 n=11+15)
MultipolygonValidation/n=4                                  1.11µs ± 4%    1.13µs ± 3%   +1.45%  (p=0.049 n=15+15)
MultipolygonValidation/n=16                                 4.50µs ± 4%    4.50µs ± 3%     ~     (p=0.717 n=15+13)
MultipolygonValidation/n=64                                 20.6µs ± 3%    20.7µs ± 3%     ~     (p=0.377 n=15+14)
MultipolygonValidation/n=256                                 134µs ± 2%     136µs ± 3%   +1.72%  (p=0.012 n=14+14)
MultipolygonValidation/n=1024                                630µs ± 3%     626µs ± 4%     ~     (p=0.486 n=15+15)
MultiPolygonTwoCircles/n=10                                 4.34µs ± 4%    4.33µs ± 2%     ~     (p=0.805 n=15+14)
MultiPolygonTwoCircles/n=100                                48.1µs ± 3%    48.2µs ± 1%     ~     (p=0.860 n=14+12)
MultiPolygonTwoCircles/n=1000                                534µs ± 4%     538µs ± 3%     ~     (p=0.254 n=13+15)
MultiPolygonTwoCircles/n=10000                              8.74ms ± 3%    8.82ms ± 4%     ~     (p=0.595 n=15+15)
MultiPolygonMultipleTouchingPoints/n=1                      6.09µs ± 3%    6.13µs ± 4%     ~     (p=0.367 n=15+15)
MultiPolygonMultipleTouchingPoints/n=10                     50.4µs ± 3%    50.5µs ± 4%     ~     (p=0.683 n=15+15)
MultiPolygonMultipleTouchingPoints/n=100                     595µs ± 3%     593µs ± 2%     ~     (p=0.550 n=13+14)
MultiPolygonMultipleTouchingPoints/n=1000                   7.30ms ± 3%    7.31ms ± 2%     ~     (p=0.458 n=14+13)
WKTParsing/point                                            2.14µs ± 3%    2.12µs ± 3%     ~     (p=0.269 n=13+15)
DistancePolygonToPolygonOrdering/n=100_swap=false           67.8µs ± 3%    68.8µs ± 5%     ~     (p=0.088 n=15+13)
DistancePolygonToPolygonOrdering/n=100_swap=true            67.7µs ± 3%    68.7µs ± 3%   +1.51%  (p=0.018 n=15+14)
DistancePolygonToPolygonOrdering/n=1000_swap=false          1.10ms ± 2%    1.10ms ± 3%     ~     (p=0.775 n=15+15)
DistancePolygonToPolygonOrdering/n=1000_swap=true           1.09ms ± 3%    1.09ms ± 2%     ~     (p=0.525 n=15+13)
IntersectionPolygonWithPolygonOrdering/n=100_swap=false     7.03µs ± 5%    7.06µs ± 4%     ~     (p=0.660 n=14+15)
IntersectionPolygonWithPolygonOrdering/n=100_swap=true      7.05µs ± 5%    6.99µs ± 3%     ~     (p=0.624 n=15+15)
IntersectionPolygonWithPolygonOrdering/n=1000_swap=false    76.7µs ± 3%    77.2µs ± 3%     ~     (p=0.252 n=15+14)
IntersectionPolygonWithPolygonOrdering/n=1000_swap=true     76.5µs ± 3%    76.3µs ± 4%     ~     (p=0.867 n=14+13)
MultiLineStringIsSimpleManyLineStrings/n=100                65.2µs ± 3%    66.3µs ± 3%   +1.63%  (p=0.016 n=15+15)
MultiLineStringIsSimpleManyLineStrings/n=1000                716µs ± 4%     730µs ± 4%   +1.97%  (p=0.012 n=14+15)
pkg:github.com/peterstace/simplefeatures/geos goos:linux goarch:amd64
IntersectionWithoutValidation/n=10                          67.8µs ±11%    66.0µs ± 4%     ~     (p=0.290 n=15+14)
IntersectionWithoutValidation/n=100                          127µs ± 4%     129µs ± 5%     ~     (p=0.290 n=15+14)
IntersectionWithoutValidation/n=1000                         589µs ± 6%     590µs ± 5%     ~     (p=0.847 n=15+14)
IntersectionWithoutValidation/n=10000                       4.87ms ± 2%    4.96ms ± 6%     ~     (p=0.131 n=12+14)
NoOp/n=10                                                   5.18µs ± 9%    5.11µs ± 4%     ~     (p=0.477 n=15+14)
NoOp/n=100                                                  16.9µs ± 4%    16.9µs ± 3%     ~     (p=0.820 n=14+12)
NoOp/n=1000                                                  123µs ± 4%     124µs ± 4%     ~     (p=0.667 n=12+14)
NoOp/n=10000                                                1.30ms ±29%    1.27ms ±21%     ~     (p=0.847 n=14+15)
pkg:github.com/peterstace/simplefeatures/internal/perf goos:linux goarch:amd64
LineStringIsSimpleCircle/n=10                               2.56µs ± 4%    2.57µs ± 3%     ~     (p=0.579 n=14+14)
LineStringIsSimpleCircle/n=100                              44.8µs ± 4%    44.2µs ± 3%     ~     (p=0.234 n=14+15)
LineStringIsSimpleCircle/n=1000                              490µs ± 2%     497µs ± 4%   +1.55%  (p=0.046 n=12+14)
LineStringIsSimpleCircle/n=10000                            6.82ms ± 3%    6.74ms ± 3%     ~     (p=0.125 n=14+14)
LineStringIsSimpleZigZag/10                                 2.35µs ± 3%    2.30µs ± 2%   -1.87%  (p=0.006 n=14+13)
LineStringIsSimpleZigZag/100                                39.5µs ± 3%    39.3µs ± 5%     ~     (p=0.451 n=14+15)
LineStringIsSimpleZigZag/1000                                463µs ± 2%     463µs ± 3%     ~     (p=0.838 n=15+15)
LineStringIsSimpleZigZag/10000                              6.71ms ± 4%    6.75ms ± 3%     ~     (p=0.505 n=14+15)
SetOperation/n=4/Go_Intersection                            57.4µs ± 5%    59.2µs ± 6%   +3.08%  (p=0.003 n=14+15)
SetOperation/n=4/Go_Difference                              58.9µs ± 2%    60.9µs ± 5%   +3.39%  (p=0.000 n=13+15)
SetOperation/n=4/Go_SymmetricDifference                     76.1µs ±11%    76.1µs ± 3%     ~     (p=0.252 n=14+15)
SetOperation/n=4/Go_Union                                   61.9µs ± 8%    62.1µs ± 4%     ~     (p=0.583 n=14+13)
SetOperation/n=4/GEOS_Intersection                          66.8µs ± 4%    67.7µs ± 8%     ~     (p=0.650 n=13+15)
SetOperation/n=4/GEOS_Difference                            70.6µs ±10%    70.8µs ± 6%     ~     (p=0.571 n=14+14)
SetOperation/n=4/GEOS_SymmetricDifference                   99.0µs ± 8%    99.2µs ± 5%     ~     (p=0.769 n=14+14)
SetOperation/n=4/GEOS_Union                                 72.0µs ±13%    70.8µs ± 8%     ~     (p=0.458 n=15+14)
SetOperation/n=8/Go_Intersection                            73.3µs ±10%    74.4µs ± 5%     ~     (p=0.068 n=14+13)
SetOperation/n=8/Go_Difference                              75.6µs ±11%    75.9µs ± 7%     ~     (p=0.488 n=13+14)
SetOperation/n=8/Go_SymmetricDifference                     95.9µs ± 5%    95.7µs ± 3%     ~     (p=0.847 n=14+15)
SetOperation/n=8/Go_Union                                   76.6µs ± 7%    77.5µs ± 3%     ~     (p=0.234 n=14+15)
SetOperation/n=8/GEOS_Intersection                          79.0µs ± 5%    78.2µs ± 9%     ~     (p=0.387 n=13+15)
SetOperation/n=8/GEOS_Difference                            79.7µs ± 4%    79.2µs ± 5%     ~     (p=0.635 n=14+14)
SetOperation/n=8/GEOS_SymmetricDifference                    118µs ±14%     116µs ± 6%     ~     (p=0.920 n=13+13)
SetOperation/n=8/GEOS_Union                                 81.9µs ±10%    82.0µs ± 7%     ~     (p=0.780 n=15+14)
SetOperation/n=16/Go_Intersection                            102µs ±14%     102µs ± 4%     ~     (p=0.482 n=14+14)
SetOperation/n=16/Go_Difference                              109µs ± 6%     108µs ± 4%     ~     (p=0.331 n=15+14)
SetOperation/n=16/Go_SymmetricDifference                     143µs ± 5%     144µs ± 6%     ~     (p=0.454 n=14+14)
SetOperation/n=16/Go_Union                                   112µs ± 5%     113µs ± 4%     ~     (p=0.202 n=15+15)
SetOperation/n=16/GEOS_Intersection                         89.1µs ±13%    86.2µs ± 9%     ~     (p=0.116 n=15+15)
SetOperation/n=16/GEOS_Difference                           95.3µs ± 7%    93.9µs ± 5%     ~     (p=0.254 n=15+13)
SetOperation/n=16/GEOS_SymmetricDifference                   147µs ± 9%     148µs ± 4%     ~     (p=0.482 n=14+14)
SetOperation/n=16/GEOS_Union                                97.7µs ± 5%    97.3µs ± 5%     ~     (p=0.847 n=15+14)
SetOperation/n=32/Go_Intersection                            173µs ± 5%     174µs ± 3%     ~     (p=0.434 n=14+11)
SetOperation/n=32/Go_Difference                              181µs ± 5%     181µs ± 2%     ~     (p=0.928 n=13+15)
SetOperation/n=32/Go_SymmetricDifference                     239µs ± 5%     239µs ± 3%     ~     (p=0.595 n=15+15)
SetOperation/n=32/Go_Union                                   188µs ± 3%     187µs ± 3%     ~     (p=0.354 n=14+15)
SetOperation/n=32/GEOS_Intersection                          107µs ±12%     104µs ± 7%     ~     (p=0.367 n=15+15)
SetOperation/n=32/GEOS_Difference                            115µs ± 4%     116µs ±11%     ~     (p=0.488 n=13+14)
SetOperation/n=32/GEOS_SymmetricDifference                   192µs ± 3%     189µs ± 6%     ~     (p=0.145 n=12+14)
SetOperation/n=32/GEOS_Union                                 124µs ± 7%     123µs ± 8%     ~     (p=0.621 n=14+15)
SetOperation/n=64/Go_Intersection                            290µs ± 6%     287µs ± 3%     ~     (p=0.300 n=15+12)
SetOperation/n=64/Go_Difference                              302µs ± 2%     307µs ± 4%   +1.45%  (p=0.035 n=12+13)
SetOperation/n=64/Go_SymmetricDifference                     414µs ± 4%     416µs ± 5%     ~     (p=0.685 n=14+13)
SetOperation/n=64/Go_Union                                   323µs ± 7%     323µs ± 5%     ~     (p=0.541 n=14+14)
SetOperation/n=64/GEOS_Intersection                          135µs ± 5%     135µs ± 5%     ~     (p=0.652 n=15+14)
SetOperation/n=64/GEOS_Difference                            165µs ± 7%     162µs ± 3%     ~     (p=0.164 n=14+14)
SetOperation/n=64/GEOS_SymmetricDifference                   287µs ± 3%     285µs ± 4%     ~     (p=0.536 n=12+14)
SetOperation/n=64/GEOS_Union                                 179µs ± 5%     177µs ± 3%     ~     (p=0.239 n=14+13)
SetOperation/n=128/Go_Intersection                           538µs ± 3%     539µs ± 4%     ~     (p=0.910 n=14+14)
SetOperation/n=128/Go_Difference                             566µs ± 5%     561µs ± 3%     ~     (p=0.323 n=15+12)
SetOperation/n=128/Go_SymmetricDifference                    738µs ± 4%     744µs ± 2%     ~     (p=0.185 n=15+13)
SetOperation/n=128/Go_Union                                  583µs ± 3%     581µs ± 7%     ~     (p=0.539 n=15+15)
SetOperation/n=128/GEOS_Intersection                         200µs ± 4%     200µs ± 5%     ~     (p=0.935 n=15+15)
SetOperation/n=128/GEOS_Difference                           241µs ± 7%     235µs ± 5%     ~     (p=0.104 n=14+14)
SetOperation/n=128/GEOS_SymmetricDifference                  450µs ± 5%     450µs ± 5%     ~     (p=0.914 n=15+14)
SetOperation/n=128/GEOS_Union                                260µs ± 7%     256µs ± 5%     ~     (p=0.454 n=14+14)
SetOperation/n=256/Go_Intersection                          1.00ms ± 5%    1.01ms ± 5%     ~     (p=0.525 n=15+13)
SetOperation/n=256/Go_Difference                            1.07ms ± 6%    1.07ms ± 4%     ~     (p=0.223 n=13+13)
SetOperation/n=256/Go_SymmetricDifference                   1.46ms ± 5%    1.46ms ± 2%     ~     (p=0.724 n=13+13)
SetOperation/n=256/Go_Union                                 1.12ms ± 6%    1.13ms ± 4%     ~     (p=0.967 n=15+15)
SetOperation/n=256/GEOS_Intersection                         292µs ± 4%     291µs ± 3%     ~     (p=0.981 n=14+13)
SetOperation/n=256/GEOS_Difference                           383µs ± 5%     379µs ± 4%     ~     (p=0.220 n=14+13)
SetOperation/n=256/GEOS_SymmetricDifference                  839µs ± 5%     846µs ± 3%     ~     (p=0.311 n=13+13)
SetOperation/n=256/GEOS_Union                                438µs ± 6%     433µs ± 3%     ~     (p=0.389 n=15+15)
SetOperation/n=512/Go_Intersection                          2.00ms ± 5%    1.99ms ± 3%     ~     (p=0.838 n=15+15)
SetOperation/n=512/Go_Difference                            2.10ms ± 4%    2.07ms ± 5%     ~     (p=0.124 n=15+15)
SetOperation/n=512/Go_SymmetricDifference                   2.83ms ± 3%    2.83ms ± 5%     ~     (p=0.667 n=14+14)
SetOperation/n=512/Go_Union                                 2.15ms ± 4%    2.16ms ± 4%     ~     (p=0.616 n=14+13)
SetOperation/n=512/GEOS_Intersection                         526µs ± 5%     539µs ± 7%   +2.34%  (p=0.044 n=14+14)
SetOperation/n=512/GEOS_Difference                           699µs ± 3%     700µs ± 7%     ~     (p=0.486 n=15+15)
SetOperation/n=512/GEOS_SymmetricDifference                 1.59ms ± 7%    1.58ms ± 5%     ~     (p=0.838 n=15+15)
SetOperation/n=512/GEOS_Union                                804µs ± 6%     808µs ±11%     ~     (p=0.838 n=15+15)
SetOperation/n=1024/Go_Intersection                         3.97ms ± 2%    4.01ms ± 4%     ~     (p=0.186 n=14+15)
SetOperation/n=1024/Go_Difference                           4.07ms ± 4%    4.14ms ± 3%   +1.87%  (p=0.036 n=12+14)
SetOperation/n=1024/Go_SymmetricDifference                  5.83ms ± 3%    5.77ms ± 3%     ~     (p=0.093 n=12+15)
SetOperation/n=1024/Go_Union                                4.32ms ± 5%    4.31ms ± 3%     ~     (p=0.983 n=15+14)
SetOperation/n=1024/GEOS_Intersection                        982µs ± 8%     994µs ± 7%     ~     (p=0.376 n=14+14)
SetOperation/n=1024/GEOS_Difference                         1.42ms ± 5%    1.44ms ± 4%     ~     (p=0.186 n=15+14)
SetOperation/n=1024/GEOS_SymmetricDifference                3.23ms ± 3%    3.21ms ± 5%     ~     (p=0.234 n=15+14)
SetOperation/n=1024/GEOS_Union                              1.68ms ± 5%    1.68ms ± 4%     ~     (p=0.880 n=14+15)
SetOperation/n=2048/Go_Intersection                         8.25ms ± 8%    8.11ms ± 3%     ~     (p=0.305 n=15+15)
SetOperation/n=2048/Go_Difference                           8.97ms ± 2%    8.85ms ± 6%   -1.34%  (p=0.023 n=12+14)
SetOperation/n=2048/Go_SymmetricDifference                  12.5ms ± 5%    12.4ms ± 4%     ~     (p=0.533 n=15+14)
SetOperation/n=2048/Go_Union                                9.12ms ± 5%    9.14ms ± 5%     ~     (p=0.838 n=15+15)
SetOperation/n=2048/GEOS_Intersection                       2.04ms ± 4%    2.04ms ± 3%     ~     (p=0.806 n=15+15)
SetOperation/n=2048/GEOS_Difference                         2.63ms ± 4%    2.61ms ± 4%     ~     (p=0.305 n=15+15)
SetOperation/n=2048/GEOS_SymmetricDifference                6.18ms ± 3%    6.17ms ± 3%     ~     (p=0.793 n=13+14)
SetOperation/n=2048/GEOS_Union                              3.10ms ± 4%    3.11ms ± 6%     ~     (p=0.967 n=15+15)
SetOperation/n=4096/Go_Intersection                         17.7ms ± 3%    17.6ms ± 5%     ~     (p=0.486 n=15+15)
SetOperation/n=4096/Go_Difference                           18.5ms ± 4%    18.2ms ± 4%     ~     (p=0.085 n=15+14)
SetOperation/n=4096/Go_SymmetricDifference                  26.1ms ± 5%    26.0ms ± 4%     ~     (p=0.683 n=15+14)
SetOperation/n=4096/Go_Union                                19.2ms ± 5%    19.1ms ± 3%     ~     (p=0.914 n=15+14)
SetOperation/n=4096/GEOS_Intersection                       3.72ms ± 4%    3.67ms ± 4%     ~     (p=0.072 n=13+13)
SetOperation/n=4096/GEOS_Difference                         5.72ms ± 4%    5.74ms ± 2%     ~     (p=0.525 n=15+13)
SetOperation/n=4096/GEOS_SymmetricDifference                14.2ms ± 5%    14.2ms ± 3%     ~     (p=0.982 n=14+14)
SetOperation/n=4096/GEOS_Union                              6.61ms ± 5%    6.62ms ± 6%     ~     (p=0.847 n=14+15)
SetOperation/n=8192/Go_Intersection                         36.9ms ± 8%    36.8ms ± 6%     ~     (p=0.867 n=13+14)
…