@peterstace 2021-02-17
I've finished self-review, ready to merge once benchmark results are in.